rework: factor out pyudev control
