+++ tests/test_usb_ctrl
@@ -1,0 +1,27 @@
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+from inet_nm.usb_ctrl import get_devices_from_tty, get_tty_from_nm_node, TtyNotPresent
+
+
+def test_get_devices_from_tty(mock_devices, dummy_nodes):
+    """Mock devices returned by context.list_devices."""
+    # Mock the context.list_devices method
+    list_devices_mock = MagicMock()
+    list_devices_mock.return_value = mock_devices
+
+    with patch("pyudev.Context.list_devices", list_devices_mock):
+        assert get_devices_from_tty() == dummy_nodes
+
+
+def test_get_tty_from_nm_node(mock_devices, dummy_nodes):
+    """Test if the tty can be checked fro a given node."""
+    with patch("pyudev.Context.list_devices", MagicMock(return_value=mock_devices[:1])):
+        assert (
+            get_tty_from_nm_node(dummy_nodes[0])
+            == mock_devices[0].device_node
+        )
+
+        with pytest.raises(TtyNotPresent):
+            get_tty_from_nm_node(dummy_nodes[1])
